requirements.txt
whitenoise
invoke
flake8
freezegun
django-toolbelt
Django==1.7.5
restless==2.0.1
six==1.9.0
Pillow==2.7.0
pycrypto==2.6.1
qrcode==5.1
python-dateutil==2.4.1
pytz
APScheduler==3.0
